pretty printing...
git-svn-id: http://svn.delph-in.net/lkb/trunk@4871 3df82f5b-d43a-0410-af33-fce91db48ec5

+++ lexdb/fns_plpgsql.sql
@@ -39,60 +39,66 @@ BEGIN
 END;
 ' LANGUAGE plpgsql;
 
+CREATE OR REPLACE FUNCTION public.index_public_revision() RETURNS boolean AS '
+BEGIN
+ 	RAISE INFO \'Creating indices...\';
+	ALTER TABLE public.revision ADD PRIMARY KEY (name,version,userid);
+	CREATE INDEX public_orthkey ON public.revision (orthkey); 
+	CREATE UNIQUE INDEX name_modstamp ON public.revision (name,modstamp); 
+	CREATE INDEX public_revision_name_modstamp ON public.revision (name, modstamp);
+	CREATE INDEX public_revision_name
+		ON public.revision (name varchar_ops); 
+	PERFORM if_version(\'7.4\', \'CREATE INDEX public_revision_name_pattern ON public.revision (name varchar_pattern_ops)\', \'CREATE INDEX public_revision_name_pattern ON public.revision (name)\');
+	RETURN true;
+END;
+' LANGUAGE plpgsql;
+
+CREATE OR REPLACE FUNCTION public.deindex_public_revision() RETURNS boolean AS '
+BEGIN
+ 	RAISE INFO \'Dropping indices...\';
+ 	DROP INDEX public_orthkey;
+ 	DROP INDEX name_modstamp;
+	DROP INDEX public_revision_name_modstamp;
+	DROP INDEX public_revision_name;
+	DROP INDEX public_revision_name_pattern;
+	ALTER TABLE public.revision DROP CONSTRAINT revision_pkey;
+	RETURN true;
+END;
+' LANGUAGE plpgsql;
+
 CREATE OR REPLACE FUNCTION public.merge_into_db2() RETURNS integer AS 
 '
 DECLARE
-	num_merged int;
+	count_new int;
 BEGIN
- PERFORM assert_db_owner();	
----- n o t e : must copy file to temp before invoking this code
-----           eg. COPY TO stdin from frontend
+	PERFORM assert_db_owner();	
+	---- n o t e : must copy file to temp before invoking this code
+	----           eg. COPY TO stdin from frontend
 
- RAISE INFO ''Preparing tables to receive data...'';
+	RAISE INFO \'Selecting new entries to merge...\';
+ 	CREATE INDEX temp_name_userid_version on temp (name, userid, version);
+ 	DELETE FROM revision_new;
+	INSERT INTO revision_new
+		SELECT * FROM (SELECT DISTINCT name,userid,version FROM temp EXCEPT SELECT name,userid,version FROM public.revision) AS t1 NATURAL JOIN temp;
+	DROP INDEX temp_name_userid_version;
+	DELETE FROM temp;
+	count_new := (SELECT count(*) FROM revision_new);
 
- DROP INDEX public_orthkey;
- DROP INDEX name_modstamp;
- DROP INDEX public_revision_name_modstamp;
- DROP INDEX public_revision_name;
- DROP INDEX public_revision_name_pattern;
- ALTER TABLE public.revision DROP CONSTRAINT revision_pkey;
+	IF count_new > 0 THEN
+ 		PERFORM public.deindex_public_revision();
 
- DELETE FROM revision_new;
+		RAISE INFO \'Inserting new % entries...\', count_new;
+		INSERT INTO public.revision SELECT * FROM revision_new;
  
- CREATE INDEX temp_name_userid_version on temp (name, userid, version);
- 
- RAISE INFO ''Copying data...'';
+ 		PERFORM public.index_public_revision();
 
- INSERT INTO revision_new
-  SELECT * FROM (SELECT DISTINCT name,userid,version FROM temp EXCEPT SELECT name,userid,version FROM public.revision) AS t1 NATURAL JOIN temp;
-
- num_merged := (SELECT count(*) FROM revision_new);
-
- RAISE INFO ''% new entries merged'', num_merged;
-
- DROP INDEX temp_name_userid_version;
- DELETE FROM temp;
- INSERT INTO public.revision SELECT * FROM revision_new;
- 
- RAISE INFO ''Recreating table relations...'';
-
- ALTER TABLE public.revision ADD PRIMARY KEY (name,version,userid);
- CREATE INDEX public_orthkey ON public.revision (orthkey); 
- CREATE UNIQUE INDEX name_modstamp ON public.revision (name,modstamp); 
- CREATE INDEX public_revision_name_modstamp ON public.revision (name, modstamp);
- CREATE INDEX public_revision_name
-  ON public.revision (name varchar_ops); 
- PERFORM 
-  if_version(''7.4'',
-   ''CREATE INDEX public_revision_name_pattern ON public.revision (name varchar_pattern_ops)'',
-   ''CREATE INDEX public_revision_name_pattern ON public.revision (name)'');
-
- RAISE INFO ''Updating timestamps...'';
-
- DELETE FROM public.meta WHERE var=''mod_time'';
- INSERT INTO public.meta VALUES (''mod_time'',current_timestamp);
-
- RETURN num_merged;
+		RAISE INFO \'Updating timestamps...\';
+		DELETE FROM public.meta WHERE var=\'mod_time\';
+		INSERT INTO public.meta VALUES (\'mod_time\',current_timestamp);
+	ELSE
+		RAISE INFO \'0 new entries\';
+	END IF;
+ RETURN count_new;
 END;
 ' LANGUAGE plpgsql;
 
@@ -101,28 +107,27 @@ CREATE OR REPLACE FUNCTION public.merge_dfn() RETURNS integer AS
 DECLARE
 	num_new int;
 BEGIN
- PERFORM assert_db_owner();	
----- n o t e : must copy file to temp_defn before invoking this code
-----           eg. COPY TO stdin from frontend
+ 	PERFORM assert_db_owner();	
+	---- n o t e : must copy file to temp_defn before invoking this code
+	----           eg. COPY TO stdin from frontend
 
- num_new := (SELECT count(*) FROM 
-             (SELECT mode, slot, field FROM temp_defn EXCEPT
-               SELECT mode, slot, field FROM defn) AS t1
-             NATURAL JOIN temp_defn);
+ 	num_new := (SELECT count(*) FROM 
+             		(SELECT mode, slot, field FROM temp_defn EXCEPT
+               			SELECT mode, slot, field FROM defn) AS t1
+             			NATURAL JOIN temp_defn);
 
- RAISE INFO ''% new field mappings'', num_new;
- RAISE INFO ''Copying data...'';
+ 	RAISE INFO \'% new field mappings\', num_new;
 
- DELETE FROM defn;
- INSERT INTO defn
-  SELECT * FROM temp_defn; 
-
- RAISE INFO ''Updating timestamps...'';
-
- DELETE FROM public.meta WHERE var=''mod_time'';
- INSERT INTO public.meta VALUES (''mod_time'',current_timestamp);
-
- RETURN num_new;
+	IF num_new > 0 THEN
+ 		RAISE INFO \'Updating table...\';
+ 		DELETE FROM defn;
+		INSERT INTO defn
+  			SELECT * FROM temp_defn; 
+ 		RAISE INFO \'Updating timestamps...\';
+ 		DELETE FROM public.meta WHERE var=\'mod_time\';
+ 		INSERT INTO public.meta VALUES (\'mod_time\',current_timestamp);
+	END IF;
+ 	RETURN num_new;
 END;
 ' LANGUAGE plpgsql;
